Update all files
