+++ .gitignore
@@ -12,6 +12,8 @@
 
 # Used by dotenv library to load environment variables.
 # .env
+#Ignore .env file
+/.env
 
 # Ignore Byebug command history file.
 .byebug_history
