+++ Gemfile
@@ -3,6 +3,12 @@ source 'https://rubygems.org'
 git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 # gem "rails"
+# source 'https://rubygems.org'
+
+git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
+
+# gem "rails"
 
 gem 'rspec'
 gem 'telegram_bot'
+gem 'dotenv'
