+++ bin/main.rb
@@ -1,4 +1,3 @@
 require_relative '../lib/bot.rb'
-require_relative '../lib/quote.rb'
 
 Bot.new
